Flytt join-funksjoner til egen fil
